General performance improvements (#256)
