+++ src/sirerrors.c
@@ -51,6 +51,38 @@ static _sir_thread_local sir_thread_err sir_te = {
     _SIR_E_NOERROR, 0, {0}, {SIR_UNKNOWN, SIR_UNKNOWN, 0} //-V616
 };
 
+#define _SIR_E_PLATFORM_ERRORFORMAT "Platform error code %d: %s"
+
+static const struct {
+    uint32_t e;
+    const char* msg;
+} sir_errors[] = {
+    {_SIR_E_NOERROR,   "The operation completed successfully"}, //-V616
+    {_SIR_E_NOTREADY,  "libsir has not been initialized"},
+    {_SIR_E_ALREADY,   "libsir is already initialized"},
+    {_SIR_E_DUPITEM,   "Item already managed by libsir"},
+    {_SIR_E_NOITEM,    "Item not managed by libsir"},
+    {_SIR_E_NOROOM,    "Maximum number of items already stored"},
+    {_SIR_E_OPTIONS,   "Option flags are invalid"},
+    {_SIR_E_LEVELS,    "Level flags are invalid"},
+    {_SIR_E_TEXTSTYLE, "Text style is invalid"},
+    {_SIR_E_STRING,    "Invalid string argument"},
+    {_SIR_E_NULLPTR,   "NULL pointer argument"},
+    {_SIR_E_INVALID,   "Invalid argument"},
+    {_SIR_E_NODEST,    "No destinations registered for level"},
+    {_SIR_E_UNAVAIL,   "Feature is disabled or unavailable"},
+    {_SIR_E_INTERNAL,  "An internal error has occurred"},
+    {_SIR_E_COLORMODE, "Color mode is invalid"},
+    {_SIR_E_TEXTATTR,  "Text attributes are invalid"},
+    {_SIR_E_TEXTCOLOR, "Text color is invalid for mode"},
+    {_SIR_E_PLUGINBAD, "Plugin module is malformed"},
+    {_SIR_E_PLUGINDAT, "Data produced by plugin is invalid"},
+    {_SIR_E_PLUGINVER, "Plugin interface version unsupported"},
+    {_SIR_E_PLUGINERR, "Plugin reported failure"},
+    {_SIR_E_PLATFORM,  _SIR_E_PLATFORM_ERRORFORMAT},
+    {_SIR_E_UNKNOWN,   "Unknown error"},
+};
+
 bool __sir_seterror(uint32_t err, const char* func, const char* file, uint32_t line) {
     if (_sir_validerror(err)) {
         sir_te.lasterror = err;
@@ -108,7 +140,7 @@ bool __sir_handleerr(int code, const char* func, const char* file, uint32_t line
         _sir_strncpy(message, SIR_MAXERROR, tmp, strnlen(tmp, SIR_MAXERROR));
 #endif
         /* cppcheck-suppress knownConditionTrueFalse */
-        if (0 == finderr && _sir_validstrnofail(message)) { //-V560
+        if (0 == finderr) { //-V547
             __sir_setoserror(code, message, func, file, line);
 #if defined(__HAVE_XSI_STRERROR_R__) || defined(__HAVE_STRERROR_S__)
         } else {
@@ -181,9 +213,10 @@ uint32_t _sir_geterror(char message[SIR_MAXERROR]) {
             }
         }
 
-        (void)snprintf(message, SIR_MAXERROR, SIR_ERRORFORMAT, sir_te.loc.func, //-V576
+        _sir_snprintf_trunc(message, SIR_MAXERROR, SIR_ERRORFORMAT, sir_te.loc.func, //-V576
             sir_te.loc.file, sir_te.loc.line, (_sir_validstrnofail(heap_msg)
-                ? heap_msg : sir_errors[_mid].msg));
+                ? heap_msg : (_sir_validstrnofail(sir_errors[_mid].msg)
+                ? sir_errors[_mid].msg : SIR_UNKNOWN)));
 
         _sir_safefree(&heap_msg);
